ch(refactor): Refactor codebase
- move the common http method to a base view

[Finishes #166950993]
