+++ src/apps/core/utilities/custom_exception_handler.py
@@ -16,8 +16,11 @@ def custom_exception_handler(exc, context):
     if response and response.status_code == status.HTTP_404_NOT_FOUND:
         response.data = {'status': 'error', 'user_message': ERRORS['USR_09']}
 
-    elif response and (response.status_code == status.HTTP_401_UNAUTHORIZED
-                       or response.status_code == status.HTTP_403_FORBIDDEN):
+    elif response and (
+            response.status_code == status.HTTP_401_UNAUTHORIZED
+            or response.status_code == status.HTTP_403_FORBIDDEN  # noqa: W503
+            or response.status_code == status.HTTP_405_METHOD_NOT_ALLOWED
+    ):  # noqa: W503
 
         response.data = {
             'status': 'error',
